+++ src/stories/dataset/DatasetErrorMockRepository.ts
@@ -10,6 +10,7 @@ import { DatasetVersionSummaryInfo } from '@/dataset/domain/models/DatasetVersio
 import { DatasetDeaccessionDTO } from '@iqss/dataverse-client-javascript'
 import { DatasetDownloadCount } from '@/dataset/domain/models/DatasetDownloadCount'
 import { CitationFormat, FormattedCitation } from '@/dataset/domain/models/DatasetCitation'
+import { DatasetTemplate } from '@/dataset/domain/models/DatasetTemplate'
 
 export class DatasetErrorMockRepository implements DatasetMockRepository {
   getAllWithCount: (
@@ -141,6 +142,14 @@ export class DatasetErrorMockRepository implements DatasetMockRepository {
     _version: string,
     _format: CitationFormat
   ) => {
+    return new Promise((_resolve, reject) => {
+      setTimeout(() => {
+        reject('Error thrown from mock')
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  getTemplates(_collectionIdOrAlias: number | string): Promise<DatasetTemplate[]> {
     return new Promise((_resolve, reject) => {
       setTimeout(() => {
         reject('Error thrown from mock')