+++ tests/e2e-integration/e2e/sections/create-dataset/CreateDataset.spec.tsx
@@ -1,5 +1,6 @@
 import { TestsUtils } from '../../../shared/TestsUtils'
 import { DatasetLabelValue } from '../../../../../src/dataset/domain/models/Dataset'
+import { DatasetHelper } from '@tests/e2e-integration/shared/datasets/DatasetHelper'
 
 const CREATE_DATASET_PAGE_URL = '/spa/datasets/root/create'
 
@@ -42,6 +43,65 @@ describe('Create Dataset', () => {
     cy.findByText(DatasetLabelValue.DRAFT).should('exist')
     cy.findByText(DatasetLabelValue.UNPUBLISHED).should('exist')
     cy.contains('Agricultural Sciences; Arts and Humanities').should('exist')
+  })
+
+  it('shows template select when a template is available and prefill fields when a template is selected', () => {
+    cy.wrap(DatasetHelper.createDatasetTemplate(), { timeout: 10000 }).then(() => {
+      cy.visit(CREATE_DATASET_PAGE_URL)
+
+      cy.wait(1000)
+
+      cy.findByTestId('dataset-template-select').should('exist').as('templateSelect')
+      cy.findByText('None').should('exist') // No default template, None is shown
+
+      cy.get('@templateSelect').within(() => {
+        cy.findByLabelText('Toggle options menu').click({ force: true })
+        cy.findByText('Dataset Template One').click({ force: true })
+      })
+
+      cy.findByLabelText(/^Title/i).should('have.value', 'Dataset Template One Title')
+
+      cy.findByText('Description')
+        .closest('.row')
+        .within(() => {
+          cy.findByLabelText(/^Text/i).should(
+            'have.value',
+            'This is the description from Dataset Template One'
+          )
+        })
+
+      cy.findByText('Subject')
+        .should('exist')
+        .closest('.row')
+        .within(() => {
+          cy.findByLabelText('Toggle options menu').click({ force: true })
+
+          cy.findByLabelText('Agricultural Sciences').should('be.checked')
+          cy.findByLabelText('Arts and Humanities').should('be.checked')
+        })
+
+      cy.findByText('Author')
+        .closest('.row')
+        .within(() => {
+          cy.findByLabelText(/^Name/i).should('have.value', 'Belicheck, Bill')
+          cy.findByLabelText(/^Identifier Type/i).should('have.value', 'ORCID')
+        })
+
+      cy.findByText(/Save Dataset/i).click()
+
+      cy.findByRole('heading', { name: 'Dataset Template One Title' }).should('exist')
+      cy.findByText(/Dataset created successfully./).should('exist')
+      cy.findByText(DatasetLabelValue.DRAFT).should('exist')
+      cy.findByText(DatasetLabelValue.UNPUBLISHED).should('exist')
+      cy.contains('Agricultural Sciences; Arts and Humanities').should('exist')
+
+      // Delete template after test
+      cy.wrap(DatasetHelper.getDatasetTemplates(), { timeout: 10000 }).then((templates) => {
+        const { id } = templates[0]
+
+        cy.wrap(DatasetHelper.deleteDatasetTemplate(id))
+      })
+    })
   })
 
   it('should redirect the user to the Login page when the user is not authenticated', () => {