+++ packages/design-system/CHANGELOG.md
@@ -12,8 +12,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
   - Add `size` prop to control the size of the button (e.g., 'sm', 'lg').
 - **DropdownButtonItem:**
   - Add `type` prop to allow specifying the type of the element.
-  - Add `target` prop for anchor elements.
-- **SelectAdvanced:** Fix word wrapping in options list to prevent overflow and ensure long text is displayed correctly.
+- **SelectAdvanced:**
+  - Fix word wrapping in options list to prevent overflow and ensure long text is displayed correctly.
+  - Support for options with a shape of `{ label: string; value: string; }[]` instead of just `string[]`.
 
 # [2.0.2](https://github.com/IQSS/dataverse-frontend/compare/@iqss/dataverse-design-system@2.0.1...@iqss/dataverse-design-system@2.0.2) (2024-06-23)